simplifications.
Created the namedtuple as a main object so the interface is visible and understood what attributes that object will expose.
Made all methods "internal" since they should not be exposed to a user and they exist only for internal consumption.
Slight simplifications of variable names.
Moved the get_bon closed to its usage so it is easier to follow.
Created a token attribute and simplified the gathering of the token by removing a method.